fix: calculate total price float64 to float32

+++ model/query/sql_query.go
@@ -22,7 +22,7 @@ type DataQueryModel struct {
 }
 
 type ResultTotalPrice struct {
-	Sum float32 `gorm:"total_price"`
+	Sum float64 `gorm:"total_price"`
 }
 
 type GroupNumberRoomByRoomType struct {
